refactor: create folder for TransactionHistory
